Initial check-in for MSCLOGGER
MSCLOGGER is a utility class in library common.
It generates 'mscgen' control files for simulations by recording all calls to and returns from TLM transport functions.
Currently, only the non-blocking interfaces of ahbctrl_test, ahbctrl and ahbmem are making use of it.
MSCLOGGER is disabled by default. It is planned to enable logging using a precompiler define (e.g. MSCLOGGER).