Ronnys Code is integrated and should proberbly work but...
SystemC 2.3.0 throws a lot of warnings to his changes in verbose.h
Moreover I did not find a way to remove this warnings without breaking
the simulation.

I've reverted to the original verbose.h you can find Ronnys changes in
git history and my attempts to make it work in verbose_cult.h

It will fail terribly with a double free corruption and does not print
any output at all.